Make more vanilla modules
