+++ package/package.json
@@ -6,6 +6,7 @@
   "main": "dist/glsl-pipeline.cjs.js",
   "module": "dist/glsl-pipeline.esm.js",
   "types": "dist/glsl-pipeline.cjs.d.ts",
+  "sideEffects": false,
   "type": "module",
   "files": [
     "dist/**",
@@ -68,6 +69,7 @@
   "devDependencies": {
     "@babel/core": "^7.23.2",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-class-properties": "^7.23.3",
     "@babel/plugin-transform-modules-commonjs": "^7.23.3",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-react": "^7.22.15",
